backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
drf-nested-routers>=0.93.4
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0
django-allauth>=0.57.0
djangorestframework-simplejwt>=5.3.0 